Update requirements
Python 3.8 to 3.11 (3.8+ is needed for Django 4.2+). Dockerfile uses Python 3.11.

Bump other dependencies, most notably psycopg2 to psycopg (version 3 but drops the tailing version number).

In the test Dockerfile, installs 3.11 as base and also install 3.8:3.12 to test the full range with tox. Currently tests still run in Docker containers and are not parallelized (tox matrix). Maybe try to switch to that as this will likely take a long time to run.